Implementando tratamento de exceção nas controllers Segurança e Cargo

+++ Service/Services/CargoService.cs
@@ -4,7 +4,9 @@ using Ecclesia.Repository.Contracts;
 using Ecclesia.Service.Contracts;
 using System;
 using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
 using System.Linq;
+using System.Net;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -21,6 +23,8 @@ namespace Ecclesia.Service.Services
 
         public async Task DeleteCargo(int id)
         {
+            var reg = _repository.GetCargo(id);
+            if (reg == null) throw new BusinessHttpResponseException(Messages.Message(HttpStatusCode.NotFound));
             await _repository.DeleteCargo(id);
         }
 
@@ -34,6 +38,7 @@ namespace Ecclesia.Service.Services
         public async Task<Cargo> GetCargo(int id)
         {
             var cargo = await _repository.GetCargo(id);
+            if (cargo ==  null) throw new BusinessHttpResponseException(HttpStatusCode.NotFound);
             return cargo;
         }
 
@@ -44,6 +49,8 @@ namespace Ecclesia.Service.Services
 
         public async Task UpdateCargo(Cargo cargo)
         {
+            var reg = _repository.GetCargo(cargo.Id);
+            if (reg == null) throw new BusinessHttpResponseException(Messages.Message(HttpStatusCode.NotFound));
             await _repository.UpdateCargo(cargo);
         }
     }
